pyproject.toml
# pyproject.toml

# -----------------------------------------------------------
# Section 1: Build System Configuration (Required)
# Defines the tools needed to build your package.
# -----------------------------------------------------------
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# -----------------------------------------------------------
# Section 2: Project Metadata (PEP 621 Standard)
# Defines the project name, version, authors, and dependencies.
# -----------------------------------------------------------
[project]
name = "orin-ai-crm"
version = "0.1.0"
description = "A short description of my Python project."
readme = "README.md"
authors = [
    {name = "Jemmy Febryan", email = "danieljemmyfebryan@gmail.com"},
]
# license = {file = "LICENSE"}
keywords = ["python", "pytest", "mocking"]

# List of required packages for your project to run
dependencies = [
    "fastapi (==0.127.0)",
    "uvicorn (==0.40.0)",
    "openai (==2.14.0)",
    "annotated-doc (==0.0.4)",
    "annotated-types (==0.7.0)",
    "anyio (==4.12.0)",
    "certifi (==2025.11.12)",
    "click (==8.3.1)",
    "distro (==1.9.0)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.11)",
    "iniconfig (==2.3.0)",
    "jinja2 (==3.1.6)",
    "jiter (==0.12.0)",
    "markupsafe (==3.0.3)",
    "numpy (==2.4.0)",
    "packaging (==25.0)",
    "pandas (==2.3.3)",
    "pluggy (==1.6.0)",
    "pydantic (==2.12.5)",
    "pydantic-core (==2.41.5)",
    "pygments (==2.19.2)",
    "pytest (==9.0.2)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.2.1)",
    "pytz (==2025.2)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "starlette (==0.50.0)",
    "tqdm (==4.67.1)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.3)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "langgraph-checkpoint-sqlite (>=3.0.1,<4.0.0)"
]

# Development/Testing dependencies (e.g., for 'pip install -e .[dev]')
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "flake8",
]

# -----------------------------------------------------------
# Section 3: Tool-Specific Configuration
# Configuration for setuptools and pytest.
# -----------------------------------------------------------

# setuptools configuration for finding packages
[tool.setuptools]
# This tells setuptools to look for packages inside the 'src' directory
# package-dir = {"" = "src"}
# This tells setuptools to find the packages automatically
packages = ["src"] 
# packages = {find = {where = ["src"]}} # Alternative style, same result

# pytest configuration (Optional)
[tool.pytest.ini_options]
minversion = "7.0"
# Configure where pytest should look for tests
testpaths = ["tests"]
# Add any default arguments for pytest runs
addopts = "--strict-markers --cov=src --cov-report=term-missing"

[tool.poetry.dependencies]
python = "3.12.3"